Implement simple HttpFileSystem handling authority
Althought in the current implementation the authority is handled in the
HttpPath, the proper way is to generate a FS for each protocol
(provider) and authority (host) to be able to work with a single base
URL as if it is a FileSystem with the root directory being the protocol
and authority.

This commit changes the implementation of HttpPath according to the
simple HttpFileSystem implementation.